+++ src/aimanager/azext_aimanager/constants.py
@@ -25,11 +25,5 @@ AI_MODEL_TABLE_TRANSFORMER = (
     "[].{Name:name, ModelId:properties.modelId, Description:properties.description}"
 )
 
-CALCULATE_COST_TABLE_TRANSFORMER = (
-    "plans[].{VmSize:vmSize, Feasible:feasible, VmsPerReplica:vmsPerReplica, "
-    "VmHourlyPrice:vmHourlyPrice, TotalHourlyPrice:totalHourlyPrice, "
-    "MaxAvailableReplicas:maxAvailableReplicas, Quantization:quantization}"
-)
-
 # Supported model source types for an AI Manager model source.
 MODEL_SOURCE_TYPES = ["HuggingFace"]